# internal/zstd: use circular buffer for backreference window