--- a/src/internal/zstd/zstd_test.go
+++ b/src/internal/zstd/zstd_test.go
@@ -120,7 +120,7 @@ func bigData(t testing.TB) []byte {
 	bigDataOnce.Do(func() {
 		bigDataBytes, bigDataErr = os.ReadFile("../../testdata/Isaac.Newton-Opticks.txt")
 		if bigDataErr == nil {
-			bigDataBytes = bytes.Repeat(bigDataBytes, 3)
+			bigDataBytes = bytes.Repeat(bigDataBytes, 20)
 		}
 	})
 	if bigDataErr != nil {